Refactor content and standards
